Split scripts to ES modules"
